add Attribute Table as Component with Lit-HTML
remove packages bootstrap

change : AttributeTable migré vers lit-html (non component)

fix : AttributeTable .substr() (dépréciée) => .slice()

fix : AttributeTable data button

change : AttributeTable migré vers lit-html (component)

change : AttributeTable migré vers lit-html (component)

fix : retrait import inutile
